Split micro depot install quote module
