+++ crates/oasis7/src/simulator/mod.rs
@@ -169,10 +169,10 @@ pub use kernel::{
     MICRO_DEPOT_PROPOSAL_EMIT_KIND, MICRO_DEPOT_UPKEEP_DATA_COST, MicroDepotActionContext,
     MicroDepotActionKind, MicroDepotConsumedResourceClass, MicroDepotDecision,
     MicroDepotDeltaClass, MicroDepotEffectPreview, MicroDepotEvalInput, MicroDepotFacilityContext,
-    MicroDepotInstallQuote, MicroDepotPlayerContext, MicroDepotPlayerFacilitySnapshot, MicroDepotPressureClass,
-    MicroDepotProposal, MicroDepotProposalResourceDebit, MicroDepotQuotePreview,
-    MicroDepotRegionContext, MicroDepotResourceDebit, MicroDepotStatus, PowerOrderFill,
-    PromptUpdateOperation, RejectReason, WorldEvent, WorldEventKind,
+    MicroDepotInstallQuote, MicroDepotPlayerContext, MicroDepotPlayerFacilitySnapshot,
+    MicroDepotPressureClass, MicroDepotProposal, MicroDepotProposalResourceDebit,
+    MicroDepotQuotePreview, MicroDepotRegionContext, MicroDepotResourceDebit, MicroDepotStatus,
+    PowerOrderFill, PromptUpdateOperation, RejectReason, WorldEvent, WorldEventKind,
     compute_micro_depot_proposal_hash, evaluate_micro_depot_quote_with_module,
     merge_kernel_rule_decisions,
 };
